graph: add alignment to the member of rte_node
The members dispatch and xstat_off of the structure rte_node
can be min cache aligned to make room for future expansion and to
make sure have better performance. Add corresponding comments.

Due to the modification of the alignment of some members of the
rte_node structure, update file release_24_11.rst.

Signed-off-by: Huichao Cai <chcchc88@163.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>