eal: fix lcore variables documentation
The lcore variables API is new in DPDK 24.11,
that's why the function rte_lcore_var_alloc() was marked experimental.
To be clearer, the whole header file (including all macros)
is marked experimental.

A change decreased the default buffer size from 1 MB to 128 kB,
missing to update the documentation, which is fixed here.

Fixes: 5bce9bed67ad ("eal: add static per-lcore memory allocation facility")
Fixes: f2fd6c2e080c ("config: limit lcore variable maximum size to 128k")

Signed-off-by: Thomas Monjalon <thomas@monjalon.net>
Signed-off-by: 0-day Robot <robot@bytheb.org>