net/bnxt/tf_ulp: performance and tuning changes for thor2 stats cache
This change adds changes the act get API to expect a physical
address for the host memory rather than a pointer to a local
virtual address. The change was made because the virt->phys
API call has a very high overhead. Also included in this change
to the mutex locks and delays within the main loop.

Fixes: 0513f0af034d ("net/bnxt/tf_ulp: add stats cache for Thor2")
Reviewed-by: Farah Smith <farah.smith@broadcom.com>
Reviewed-by: Jay Ding <jay.ding@broadcom.com>
Signed-off-by: Peter Spreadborough <peter.spreadborough@broadcom.com>
Signed-off-by: Sriharsha Basavapatna <sriharsha.basavapatna@broadcom.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>