ethdev: allow using devices with shortened PCI addresses
When using a secondary process, e.g. pdump, not all PCI devices will be
probed correctly if the primary process has allowlisted devices using
the shortened syntax, omitting the initial "0000:" domain. This is due
to the strcmp failing to match a full address with a shortened one.

Fix this issue by comparing the device names/addresses normally and then
having a special case second comparison for devices which start with
"0000:".

Fixes: d948f596fee2 ("ethdev: fix port data mismatched in multiple process model")
Cc: stable@dpdk.org

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>