net/mlx5: fix the leak of action data list
In the actions construction for HWS non-template API, the `masks`
parameter is always set to NULL and all the actions will be
translated in the "construct" stage as non-fixed ones.

In the stage of translating actions template, the actions data would
be allocated from the pool and managed in a list. The list would be
released when destroying the template with the actions. In the NT2HWS
implementation, the temporary template was freed directly and the
actions will be destroyed with the flow rule deletion. No other rule
would use this list anymore.

The actions data in the list should be freed when the actions
construction is done.

Fixes: ff4064d5b1fe ("net/mlx5: support bulk actions in non-template flow")
Cc: mkashani@nvidia.com

Signed-off-by: Bing Zhao <bingz@nvidia.com>
Acked-by: Dariusz Sosnowski <dsosnowski@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>