examples/l3fwd: add option to set Tx burst size
The application send packets only when the buffer is full, or the
buffer is empty and the packets to be sent extends TX_PKT_BURST.
The change of MAX_PKT_BURST make TX buffer size and TX_PKT_BURST
increase, while the default cache size is 256. The packets in
the TX direction occupy the cache. As a result, the performance
deteriorates.

Restore the default Tx burst and add option '--tx-burst' to set
the Tx burst size. To ensure consistency, rename the option
'--burst' to '--rx-burst'. The valid range of the user-provided
value is (0, MAX_PKT_BURST] for both directions.

Fixes: d5c4897ecfb2 ("examples/l3fwd: add option to set RX burst size")
Cc: stable@dpdk.org

Signed-off-by: Jie Hai <haijie1@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>