net/ice/base: support switch recipe reuse
If FW supports the corresponding functionality, the driver allows PFs
to subscribe the same switch recipes. Then when the PF is done
with a switch recipes, the PF can ask the FW to free that switch recipe.

When the FW detects that all subscribing PFs have freed the switch
recipe, the FW will free the switch recipe so that it can be reused.

This feature also fixes a problem where all switch recipes would eventually
be exhausted because switch recipes could not be freed, as freeing a shared
recipe could potentially break other PFs that were using it.

This patch synchronize with the share code in kernel driver to support
reusing the switch recipe.

Signed-off-by: Zhichao Zeng <zhichaox.zeng@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>