dts: bug fixes for paramiko timeout and buffered_pipe issues
There exists some inconsistencies in the successful initialization of
interactive applications within remote sessions. On some testbeds, there
is a consistent failure in the first start-up attempt of an application,
and 'stdin' and 'stdout' buffered pipes also indicate improper
deallocation.

The issue stems from a remote server's inability to fully exit out of an
interactive application before the complete clean up of the
SingleInteractiveShell context manager. Moreover, the experience of
buffered pipe or related exceptions as a result of 'stdin' and 'stdout'
buffered pipes is a result of the context manager's cleanup process.
Explicit calls to close the 'stdin' buffered pipe is needless and this
is done via a 'Channel.close()' within Paramiko. The following patch
corrects these issues.

Bugzilla ID: 1597

Signed-off-by: Nicholas Pratte <npratte@iol.unh.edu>
Signed-off-by: 0-day Robot <robot@bytheb.org>