drivers/common: avoid truncation of constant value
This issue was flagged by MSVC warning below:

drivers\common\idpf\base/virtchnl2.h(269): warning C4309:
    'initializing': truncation of constant value

The problem is that 64-bit numbers are initialized in an enum.

The C11 standard states: The expression that defines the value of an
enumeration constant shall be an integer constant expression that
has a value representable as an int. [Section 6.7.2.2]

As a result compilers end up truncating these numbers. MSVC and Clang
do that. Gcc has an extension that makes it use a 64-bit size for the
enum to avoid the truncation.

At the moment this truncation is only a real problem for
VIRTCHNL2_CAP_OEM, but clearly all these capability flags are
intended to be 64 bits long and therefore should better not be defined
in an enum.

The fix is to use "defines" for them, as the code was using about
7 months ago before enums were introduced here. In this patch the
flag names and values are being preserved, only the enum is being
removed.

Signed-off-by: Andre Muezerie <andremue@linux.microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>