app/test-cmdline: add test to Windows build
Header termios.h was not found on Windows, which was probably the
reason this test was originally excluded from the Windows build.
However, ir turns out this header is not needed for this test, so it
can be removed and then test can be added to the Windows build
without negative consequences.

Signed-off-by: Andre Muezerie <andremue@linux.microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>