ethdev: clarify do not something in the new event
If application verify the validity of the port id or configure this port in
the new event callback, application may happen to the port id is invalid.

Actually, when application receive a new event from one port, the port is
not fully probed and is just in allocated state. Application doesn't need
to verify the validity of the port id because it is definitely valid.
What's more, application shouldn't do something like configuring this port
or querying some information of this port by ethdev ops.

Signed-off-by: Huisong Li <lihuisong@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>