eventdev: add atomic queue to test-eventdev app
Add an atomic queue test based on the order queue test that exclusively uses atomic queues.
This makes it compatible with event devices such as the distributed software eventdev.

The test detects if port maintenance is required.

To verify atomicity, a spinlock is set up for each combination of queue and flow.
It is taken whenever an event is dequeued for processing and released when processing is finished.
The test will fail if a port attempts to take a lock which is already taken.

Signed-off-by: Luka Jankovic <luka.jankovic@ericsson.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>