dts: port over queue start/stop suite
This suite tests the ability of the Poll Mode Driver
to enable and disable Rx/Tx queues on a port.
The verify argument in the deferred start method has been excluded
due to a bug in testpmd, which renders checking port queue info
for deferred start status impossible within a single method.

Signed-off-by: Dean Marx <dmarx@iol.unh.edu>
Reviewed-by: Jeremy Spewock <jspewock@iol.unh.edu>
Signed-off-by: 0-day Robot <robot@bytheb.org>