net/af_packet: provide packet drop stats
The Linux kernel provides the ability to query the packet drop counter
of a socket. This information can be provided when the user requests
stats.

It is important to note that each call to getsockopt with
PACKET_STATISTICS resets the internal counters. So the caller needs to
keep track of the total count on its own.

Next, I have added a counter for the case when mbuf couldn't be
allocated.

Signed-off-by: Stefan Laesser <stefan.laesser@omicronenergy.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>