app/testpmd: add attach and detach port for multiple process
The port information needs to be updated due to attaching and detaching
port. Currently, it is done in the same thread as removing or probing
device, which doesn't satisfy the operation of attaching and detaching
device in multiple process.

If this operation is performed in one process, the other process can
receive 'new' or 'destroy' event. So we can move updating port information
to event callback to support attaching and detaching port in primary and
secondary process.

Note: the reason for adding an alarm callback in 'destroy' event is that
the ethdev state is changed from 'ATTACHED' to 'UNUSED' only after the
event callback finished. But the remove_invalid_ports() function removes
invalid port only if ethdev state is 'UNUSED'. If we don't add alarm
callback, this detached port information can not be removed.

Signed-off-by: Huisong Li <lihuisong@huawei.com>
Signed-off-by: Dongdong Liu <liudongdong3@huawei.com>
Acked-by: Chengwen Feng <fengchengwen@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>