build: allow disabling avx512 support via compiler flag
DPDK build checks for build support for various instruction sets by
checking both the target machine and the compiler for support. However,
any disabling of instruction sets via compiler flags was not taken into
account in many cases. For AVX512 support, check for a user-specified
"no-avx512f" flag in the c_args parameter before checking if the
compiler can actually build AVX512 code.

As well as providing an option for the user, this can be used to test
builds without AVX512 without having to use an older compiler.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>