ethdev: add new API to enable/disable xstat by counters ID
This patch introduces a new API to dynamically enable or disable
xstat counters by their IDs.
Some counters may require resources that are potentially limited,
so providing the ability to toggle them on or off makes sense.
In addition, adding an API to query the current state
(enabled, disabled, or not supported) of an xstat counter is added.

New APIs:
	- rte_eth_xstats_set_counter
		(struct rte_eth_dev *dev, uint64_t id, int on_off);
 	- rte_eth_xstats_query_state
		(struct rte_eth_dev *dev, uint64_t id);

Note that by default, generic stats (like those provided by
eth_basic_stats_get()) will be considered unsupported and
will not be toggleable.
Also all xstats will be considered unsupported for dynamic enable/disable,
and each PMD will be able to override this in its implementation.

Signed-off-by: Shani Peretz <shperetz@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>